add snowflake hook, rename callable file because it interfered with imports, add create tables statement to dag
